# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Тунисский формат tournament service
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterBody {
  email: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: string;
  email: string;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface AdminUser {
  id: string;
  email: string;
  role: string;
  createdAt: string;
  tournamentCount: number;
}

export interface Player {
  id: number;
  name: string;
  gamesPlayed: number;
  wins: number;
  losses: number;
  pointsDiff: number;
}

export interface Round {
  round: number;
  restingPlayerId: number;
  teamA: number[];
  teamB: number[];
  /** @nullable */
  scoreA: number | null;
  /** @nullable */
  scoreB: number | null;
  /** @nullable */
  winner: string | null;
  completed: boolean;
  manuallyEditedTeams: boolean;
}

export interface Tournament {
  id: string;
  userId: string;
  createdAt: string;
  targetScore: number;
  players: Player[];
  rounds: Round[];
  status: string;
  /** @nullable */
  finishedAt: string | null;
}

export interface TournamentSummary {
  id: string;
  userId: string;
  createdAt: string;
  targetScore: number;
  status: string;
  /** @nullable */
  finishedAt: string | null;
  playerNames: string[];
  completedRounds: number;
}

export interface CreateTournamentBody {
  targetScore: number;
  playerNames: string[];
}

export interface UpdateRoundBody {
  /** @nullable */
  scoreA?: number | null;
  /** @nullable */
  scoreB?: number | null;
  /** @nullable */
  teamA?: number[] | null;
  /** @nullable */
  teamB?: number[] | null;
}