# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Тунисский формат tournament service
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: tournaments
    description: Tournament operations
  - name: rounds
    description: Round operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tournaments:
    get:
      operationId: listTournaments
      tags: [tournaments]
      summary: List tournaments for current user
      responses:
        "200":
          description: List of tournaments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TournamentSummary"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createTournament
      tags: [tournaments]
      summary: Create a new tournament
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTournamentBody"
      responses:
        "201":
          description: Tournament created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tournaments/{id}:
    get:
      operationId: getTournament
      tags: [tournaments]
      summary: Get tournament by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tournament
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTournament
      tags: [tournaments]
      summary: Delete tournament
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tournaments/{id}/rounds/{roundNumber}:
    patch:
      operationId: updateRound
      tags: [rounds]
      summary: Update round score or teams
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: roundNumber
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoundBody"
      responses:
        "200":
          description: Round updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (superadmin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/users/{userId}:
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete user and all their data (superadmin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/users/{userId}/tournaments:
    get:
      operationId: adminListUserTournaments
      tags: [admin]
      summary: List all tournaments for a user (superadmin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of tournaments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TournamentSummary"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token

    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string
      required:
        - id
        - email
        - role
        - createdAt

    AdminUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string
        tournamentCount:
          type: integer
      required:
        - id
        - email
        - role
        - createdAt
        - tournamentCount

    Player:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        gamesPlayed:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        pointsDiff:
          type: integer
      required:
        - id
        - name
        - gamesPlayed
        - wins
        - losses
        - pointsDiff

    Round:
      type: object
      properties:
        round:
          type: integer
        restingPlayerId:
          type: integer
        teamA:
          type: array
          items:
            type: integer
        teamB:
          type: array
          items:
            type: integer
        scoreA:
          type: ["integer", "null"]
        scoreB:
          type: ["integer", "null"]
        winner:
          type: ["string", "null"]
        completed:
          type: boolean
        manuallyEditedTeams:
          type: boolean
      required:
        - round
        - restingPlayerId
        - teamA
        - teamB
        - scoreA
        - scoreB
        - winner
        - completed
        - manuallyEditedTeams

    Tournament:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        createdAt:
          type: string
        targetScore:
          type: integer
        players:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        rounds:
          type: array
          items:
            $ref: "#/components/schemas/Round"
        status:
          type: string
        finishedAt:
          type: ["string", "null"]
      required:
        - id
        - userId
        - createdAt
        - targetScore
        - players
        - rounds
        - status
        - finishedAt

    TournamentSummary:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        createdAt:
          type: string
        targetScore:
          type: integer
        status:
          type: string
        finishedAt:
          type: ["string", "null"]
        playerNames:
          type: array
          items:
            type: string
        completedRounds:
          type: integer
      required:
        - id
        - userId
        - createdAt
        - targetScore
        - status
        - finishedAt
        - playerNames
        - completedRounds

    CreateTournamentBody:
      type: object
      properties:
        targetScore:
          type: integer
        playerNames:
          type: array
          items:
            type: string
      required:
        - targetScore
        - playerNames

    UpdateRoundBody:
      type: object
      properties:
        scoreA:
          type: ["integer", "null"]
        scoreB:
          type: ["integer", "null"]
        teamA:
          type: ["array", "null"]
          items:
            type: integer
        teamB:
          type: ["array", "null"]
          items:
            type: integer