# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Тунисский формат tournament service
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  id: number;
  name: string;
  gamesPlayed: number;
  wins: number;
  losses: number;
  pointsDiff: number;
}