# lib/api-zod/src/generated/types/round.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Тунисский формат tournament service
 * OpenAPI spec version: 0.1.0
 */

export interface Round {
  round: number;
  restingPlayerId: number;
  teamA: number[];
  teamB: number[];
  /** @nullable */
  scoreA: number | null;
  /** @nullable */
  scoreB: number | null;
  /** @nullable */
  winner: string | null;
  completed: boolean;
  manuallyEditedTeams: boolean;
}