# lib/api-zod/src/generated/types/tournament.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Тунисский формат tournament service
 * OpenAPI spec version: 0.1.0
 */
import type { Player } from "./player";
import type { Round } from "./round";

export interface Tournament {
  id: string;
  userId: string;
  createdAt: string;
  targetScore: number;
  players: Player[];
  rounds: Round[];
  status: string;
  /** @nullable */
  finishedAt: string | null;
}