# lib/api-zod/src/generated/types/updateRoundBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Тунисский формат tournament service
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateRoundBody {
  /** @nullable */
  scoreA?: number | null;
  /** @nullable */
  scoreB?: number | null;
  /** @nullable */
  teamA?: number[] | null;
  /** @nullable */
  teamB?: number[] | null;
}